add test to create appointment request
